+++ MapGen/Source/MapGen/Pos.h
@@ -4,13 +4,13 @@
 
 class MAPGEN_API Pos {
 public:
-	int32 x;
-	int32 y;
+	float x;
+	float y;
 
-	Pos(int32 x = 0, int32 y = 0);
+	Pos(float x = 0, float y = 0);
 	Pos(FCoord coord);
 
 	bool operator==(const Pos p);
 
-	int32 DistanceTo(Pos target);
+	float DistanceTo(Pos target);
 };
